+++ src/util/config.tsx
@@ -29,7 +29,6 @@ export function createBaseConfig(): SurrealistConfig {
 		lastPromptedVersion: null,
 		tabSearch: false,
 		defaultDesignerNodeMode: 'fields',
-		defaultDesignerLayoutMode: 'diagram',
-		combineResults: false
+		defaultDesignerLayoutMode: 'diagram'
 	};
 }